Bug 1450025 - Convert associated services to helper and model
