+++ package.json
@@ -19,7 +19,6 @@
     "angular": "1.7.2",
     "angular-clipboard": "1.6.2",
     "angular-local-storage": "0.7.1",
-    "angular-marked": "1.2.2",
     "angular-resource": "1.7.2",
     "angular-route": "1.7.2",
     "angular-sanitize": "1.7.2",
