+++ ui/entry-index.js
@@ -30,14 +30,11 @@ import './job-view/JobView';
 import './js/directives/treeherder/main';
 import './js/services/main';
 import './js/services/buildapi';
-import './js/services/taskcluster';
 import './js/services/classifications';
 import './js/services/jobfilters';
-import './js/services/tcactions';
 import './js/models/resultset';
 import './js/models/resultsets_store';
 import './js/models/repository';
 import './js/models/perf/series';
 import './js/controllers/main';
-import './js/controllers/tcjobactions';
 import './js/filters';
