+++ ui/entry-logviewer.js
@@ -19,8 +19,6 @@ import './js/directives/treeherder/log_viewer_steps';
 import './js/directives/treeherder/main';
 import './js/components/logviewer/logviewer';
 import './js/services/main';
-import './js/services/taskcluster';
 import './js/models/resultset';
-import './js/services/tcactions';
 import './js/filters';
 import './js/controllers/logviewer';
