+++ ui/job-view/details/summary/ActionBar.jsx
@@ -4,13 +4,14 @@ import { Queue, slugid } from 'taskcluster-client-web';
 import $ from 'jquery';
 import jsyaml from 'js-yaml';
 
-import thTaskcluster from '../../../js/services/taskcluster';
 import { thEvents } from '../../../js/constants';
 import { formatModelError, formatTaskclusterError } from '../../../helpers/errorMessage';
 import { isReftest } from '../../../helpers/job';
+import taskcluster from '../../../helpers/taskcluster';
 import { getInspectTaskUrl, getReftestUrl } from '../../../helpers/url';
 import JobDetailModel from '../../../models/jobDetail';
 import JobModel from '../../../models/job';
+import TaskclusterModel from '../../../models/taskcluster';
 import CustomJobActions from '../../CustomJobActions';
 import LogUrls from './LogUrls';
 
@@ -23,11 +24,12 @@ export default class ActionBar extends React.Component {
     this.thNotify = $injector.get('thNotify');
     this.thBuildApi = $injector.get('thBuildApi');
     this.ThResultSetStore = $injector.get('ThResultSetStore');
-    this.tcactions = $injector.get('tcactions');
     this.$interpolate = $injector.get('$interpolate');
     this.$uibModal = $injector.get('$uibModal');
     this.$rootScope = $injector.get('$rootScope');
     this.$timeout = $injector.get('$timeout');
+
+    this.taskclusterModel = new TaskclusterModel(this.thNotify);
 
     this.state = {
       customJobActionsShowing: false,
@@ -130,13 +132,13 @@ export default class ActionBar extends React.Component {
     if (selectedJob.build_system_type === 'taskcluster' || selectedJob.reason.startsWith('Created by BBB for task')) {
       this.ThResultSetStore.getGeckoDecisionTaskId(
         selectedJob.result_set_id).then(decisionTaskId => (
-          this.tcactions.load(decisionTaskId, selectedJob).then((results) => {
+          this.taskclusterModel.load(decisionTaskId, selectedJob).then((results) => {
             const actionTaskId = slugid();
             if (results) {
               const backfilltask = results.actions.find(result => result.name === 'backfill');
               // We'll fall back to actions.yaml if this isn't true
               if (backfilltask) {
-                return this.tcactions.submit({
+                return this.taskclusterModel.submit({
                   action: backfilltask,
                   actionTaskId,
                   decisionTaskId,
@@ -162,7 +164,7 @@ export default class ActionBar extends React.Component {
             }
 
             // Otherwise we'll figure things out with actions.yml
-            const queue = new Queue({ credentialAgent: thTaskcluster.getAgent() });
+            const queue = new Queue({ credentialAgent: taskcluster.getAgent() });
 
             // buildUrl is documented at
             // https://github.com/taskcluster/taskcluster-client-web#construct-urls
@@ -183,7 +185,7 @@ export default class ActionBar extends React.Component {
                 action_args: `--project=${repoName}' --job=${selectedJob.id}`,
               });
 
-              const task = thTaskcluster.refreshTimestamps(jsyaml.safeLoad(action));
+              const task = taskcluster.refreshTimestamps(jsyaml.safeLoad(action));
               queue.createTask(actionTaskId, task).then(function () {
                 this.$timeout(() => this.thNotify.send(
                   `Request sent to backfill job via actions.yml (${actionTaskId})`,
@@ -386,7 +388,6 @@ export default class ActionBar extends React.Component {
           job={selectedJob}
           pushId={selectedJob.result_set_id}
           isLoggedIn={user.isLoggedIn}
-          tcactions={this.tcactions}
           notify={this.thNotify}
           toggle={this.toggleCustomJobActions}
         />}
