+++ ui/js/auth/AuthService.js
@@ -1,5 +1,5 @@
 import { userSessionFromAuthResult, renew, loggedOutUser } from './auth-utils';
-import thTaskcluster from '../services/taskcluster';
+import taskcluster from '../../helpers/taskcluster';
 import { getApiUrl } from '../../helpers/url';
 import UserModel from '../../models/user';
 
@@ -98,6 +98,6 @@ export default class AuthService {
     localStorage.setItem('userSession', JSON.stringify(userSession));
     localStorage.setItem('user', JSON.stringify(user));
 
-    thTaskcluster.updateAgent();
+    taskcluster.updateAgent();
   }
 }
